PHASE 1/requirements_phase1.txt
groq>=0.12.0
google-play-scraper>=1.2.7
pandas>=2.1.0
python-dotenv>=1.0.0
streamlit>=1.38.0
markdown2>=2.5.0
